Refactor: 단일 책임 원칙을 기반으로 함수 분할

+++ src/main/java/com/bbangle/bbangle/search/dto/response/SearchBoardResponse.java
@@ -7,14 +7,14 @@ import lombok.Builder;
 @Builder
 public record SearchBoardResponse(
     List<BoardResponseDto> content,
-    int itemAllCount,
+    Long itemAllCount,
     int limitItemCount,
     int currentItemCount,
     int pageNumber,
     boolean existNextPage
 ) {
 
-    public static SearchBoardResponse getEmpty(int pageNumber, int limitItemCount, int itemAllCount){
+    public static SearchBoardResponse getEmpty(int pageNumber, int limitItemCount, Long itemAllCount){
         return SearchBoardResponse.builder().content(List.of())
                 .itemAllCount(itemAllCount)
                 .pageNumber(pageNumber)
